+++ sample/MinimalOpenAPI.SmokeTest.Api/Program.cs
@@ -1,5 +1,3 @@
-using Microsoft.Extensions.FileProviders;
-
 using MinimalOpenAPI;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -9,15 +7,6 @@ builder.Services.AddMinimalOpenApi();
 var app = builder.Build();
 
 app.MapMinimalOpenApiEndpoints();
-
-var openApiDirectory = Path.Combine(AppContext.BaseDirectory, "openapi");
-if (Directory.Exists(openApiDirectory))
-{
-    app.UseStaticFiles(new StaticFileOptions
-    {
-        FileProvider = new PhysicalFileProvider(openApiDirectory),
-        RequestPath = "/openapi",
-    });
-}
+app.MapOpenApiSchemas();
 
 app.Run();